+++ src/MessageHandler/AddBookingToCacheHandler.php
@@ -2,7 +2,10 @@
 
 namespace App\MessageHandler;
 
+use App\Entity\Resources\AAKResource;
+use App\Enum\UserBookingStatusEnum;
 use App\Message\AddBookingToCacheMessage;
+use App\Repository\Resources\AAKResourceRepository;
 use App\Service\BookingServiceInterface;
 use App\Service\UserBookingCacheServiceInterface;
 use Psr\Log\LoggerInterface;
@@ -19,6 +22,7 @@ class AddBookingToCacheHandler
         private readonly BookingServiceInterface $bookingService,
         private readonly UserBookingCacheServiceInterface $userBookingCacheService,
         private readonly LoggerInterface $logger,
+        private readonly AAKResourceRepository $resourceRepository,
     ) {
     }
 
@@ -32,6 +36,17 @@ class AddBookingToCacheHandler
         $booking = $message->getBooking();
         $id = $this->bookingService->getBookingIdFromICalUid($message->getICalUID()) ?? null;
 
+        $resourceEmail = $booking->getResourceEmail();
+
+        $resourceDisplayName = $booking->getResourceName();
+
+        /** @var AAKResource $resource */
+        $resource = $this->resourceRepository->findOneBy(['resourceMail' => $resourceEmail]);
+
+        if (null != $resource && $resource->getResourceDisplayName()) {
+            $resourceDisplayName = $resource->getResourceDisplayName();
+        }
+
         if (null != $id) {
             $this->userBookingCacheService->addCacheEntryFromArray([
                 'subject' => $booking->getSubject(),
@@ -39,8 +54,9 @@ class AddBookingToCacheHandler
                 'body' => $booking->getBody(),
                 'start' => $booking->getStartTime(),
                 'end' => $booking->getEndTime(),
-                'status' => 'AWAITING_APPROVAL',
+                'status' => UserBookingStatusEnum::AWAITING_APPROVAL->name,
                 'resourceMail' => $booking->getResourceEmail(),
+                'resourceDisplayName' => $resourceDisplayName,
             ]);
         } else {
             throw new RecoverableMessageHandlingException(sprintf('Booking id could not be retrieved for booking with iCalUID: %s', $message->getICalUID()));